Combat Refactor
Make a wrapper implementation of IController that takes a CombatBoard, rather than CombatBoard being an IController itself.

Have CombatBoard be the sole componant that manipulates the interfaces - other objects can do it vicariously through methods in CombatBoard such as UpdateStatsFor.

Add the methods PrepareForMove and PostMove to the IControllerInterface and have them also called in the Run cooroutine in CombatCoordinator to allow the controller with the board to lock and unlock it is neccessary and highlight certain units.